refactor: Upgrade All Packages and Replace Babel with TypeScript
In retrospect, we should have done the Babel -> TypeScript change
separately since there are always gotchas that come with package
updates. But it's not worth perfecting the Git History right now,
so we'll just commit as is. The changes aren't too complicated.

For the most part, the code stayed the same. Primary Changes:
- Replace Babel with TypeScript (for simpler configuration).
  - Fix types in all files because of strict TS rules.
- Upgrade all packages.
  - Explicitly convert WebSocket message data to `string`s after new upgrade.
- Use Node's native ESModules.
  - Replace Jest with Vitest because of ESModules.
- Update GitHub username in License.
- Warn about new experiments in README.
  NOTE: The README "article" wasn't updated and probably won't be updated until
  we finish our experiments.